+++ synphot/units.py
@@ -144,9 +144,9 @@ def convert_flux(wavelengths, fluxes, out_flux_unit, **kwargs):
         OBMAG and count, otherwise it is not needed.
 
     vegaspec : `~synphot.spectrum.SourceSpectrum`
-        Vega spectrum that *must* be provided for conversions involving
-        VEGAMAG, otherwise it is not needed. For instance, it can be
-        obtained from :func:`~synphot.spectrum.SourceSpectrum.from_vega`.
+        If this kwarg ispresent, the given Vega spectrum to use for conversion
+        of VEGAMAG. If not present, it is automatically loaded from
+        :func:`~synphot.spectrum.SourceSpectrum.from_vega`.
 
     Returns
     -------
@@ -223,6 +223,10 @@ def _convert_flux(wavelengths, fluxes, out_flux_unit, area=None,
     if VEGAMAG.to_string() in flux_unit_names:
         from synphot.spectrum import SourceSpectrum
 
+        if vegaspec is None:
+            vegaspec = SourceSpectrum.from_vega()
+
+        # Vega data file not found, so `from_vega` returns None
         if not isinstance(vegaspec, SourceSpectrum):
             raise exceptions.SynphotError('Vega spectrum is missing.')